A document named Testing
Required under FinalProjectSubmission.pdf
 
"System and Unit tests. Include a document under GIT called Testing. Include in this document a list of the tests you have performed for each unit and for the entire system. For acceptance test, you will choose a number of tests from the system test to demonstrate the functionality of the system. Automated build and unit tests. "